refactor: move Bytes formatting to bytes.{h,cc}
This makes it so that `Value::op<<` can just stream the `Bytes`
directly, and `Value` is no longer responsible for the
formatting/escaping of bytes.

This also makes it so that if a user has a `Bytes` object (not a
`Value`), they can still stream it.

There's no change in formatting logic or behavior in this PR. Just a
refactoring/move of the code from value.cc -> bytes.cc.