refactor: move Bytes formatting to bytes.{h,cc} (#1384)
This makes it so that Value::op<< can just stream the Bytes
directly, and Value is no longer responsible for the
formatting/escaping of bytes.

This makes it so that if a user has a Bytes object (not a
Value), they can still stream it. It's also more efficient this way because we can
directly iterate using the Bytes::Decoder rather than having to first copy the
bytes to an unnecessary vector<unsigned char>.

There's no change in formatting logic or behavior in this PR. Just a
refactoring/move of the code from value.cc -> bytes.cc.