feat: add async calls to `SessionPool` (#1390)
* feat: add async calls to `SessionPool`

note #1389 is effectively a DIFFBASE; you only need to review the
changes in `session_pool.{h,cc}` in this PR.

Adds wrappers for Async{BatchCreate,Delete,Get}Session that use the
retry loops from #1389 and returns a future for the result of the call.

These could undoubtably use some tests.

fixes #1365

* address review comments