feat: clean up main doxygen page
Much of the getting-started and compiling stuff I'm not sending the user
to the quickstart/ repo, which has code and instructions that we're
testing, and I think it's nice and simple.

Moved the info about authentication into the quickstart/README.md.

Added a summary of most of the environment variables that can configure
the library.
Fixes: #1306