feat: implement async session creation (#1405)
* feat: implement async session creation

Also enable session creation in the background thread.

I'm planning to change the existing uses of synchronous creation to
async but that requires very substantial test changes. However, I did
run a couple experiments - one *only* using `AsyncBatchCreateSessions()`,
and one with that, and *only* calling `Grow()` from the background
thread (that too some hackery) and all of the integration tests,
benchmarks, and samples passed in both cases. So, I wanted to get this
out for review while I work on making the unit tests function with async
ops.

Part of #1271

* address review comments

* fix windows compile error