fix: remove `includes = ["."]`
Using `includes = ["."]` in a top-level BUILD file causes too many odd
behaviors and breakages. For example, it makes `bazel build ...` fail.

The goal was to allow users to include our headers using angle-bracket
includes, which this does. However, there may be a better way to do
this. Until we figure that out, we should remove the odd line.