feat: implement SessionPool session refresh
Call `GetSession()` during `SessionPool::DoBackgroundWork()` for any
session whose last-use time is older than the keep-alive interval.

Also change the type of `SessionPoolOptions::keep_alive_interval_`
from `std::chrono::minutes` to `std::chrono::seconds`.  This allows
the running time of the test to be drastically reduced. It is not
an API breakage because `minutes` convert implicitly to `seconds`.

Fixes #1171.