feat!: implement SessionPool session refresh (#1425)
Call `GetSession()` during `SessionPool::DoBackgroundWork()` for any
session whose last-use time is older than the keep-alive interval.

Also change the type of `SessionPoolOptions::keep_alive_interval_`
from `std::chrono::minutes` to `std::chrono::seconds`.  This is API
compatible for callers of `SessionPoolOptions::set_keep_alive_interval()`,
but is a breakage if the result of `SessionPoolOptions::keep_alive_interval()`
is assigned to a `std::chrono::minutes`.  We believe this is unlikely to
impact many/any users.

Fixes #1171.