test: write integration test for SessionPool (#1442)
Make sure the low-level functions to make async calls actually work,
the value of unit tests for these is limited (we would be re-testing
the retry loop, but we have tests for that). An integration test
verifies we are filling the protos with the right data. Note that
some of these functions are not exercised in the other integration
tests because they run too infrequently.