fix: made .clang-format generic for any repo
As we prepare to move to a monorepo, we want all of our existing repos
to be formatted with the same .clang-format config. This config removes
any per-library specific names (i.e., I removed "spanner"). Instead,
this names the headers and directories of the *common* headers and
ensures that those sort after the other quoted "google/cloud/.+"
includes, which should be the project ones.