ci: avoid modifying the developer's gcloud config
To upload and download the cache we need to add a service account to the
developer's environment. For the CI builds this does not matter, as the
CI runs with a clean environment. When running the builds locally this
was annoying because it revoked the developer's credentials.

These changes fix that, probably using more changes than strictily
necessary, but it was easier to copy the code from g-c-cpp.