cleanup: make unit test target names unique (#1471)
CMake requires target names to be unique, even in different directories.
Fortunately the target name can be different from the binary name,
though it requires a little more work to make it so. Unit tests are more
likely to repeat names, so use a prefix to minimize the possibility of
collisions. The other targets can be handled on a ad-hoc basis when and
if we have a target name collision.