chore: link with gtest_main (#1472)
* chore: link with gtest_main

Fixes https://github.com/googleapis/google-cloud-cpp/issues/3713

Currently, the `:google_cloud_cpp_testing` target _also_ includes a
`main()`, that is exactly the same as `:gtest_main`. However, that
target is losing its implementation of main (it's gone in `-cpp`
already). So to keep the `-spanner` code working when it moves into
`-cpp`, it should explicitly link with `:gtest_main` now.

* removed use of init_google_mock.h